requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Pillow==10.0.0
pymysql==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
torch==2.0.1
timm==0.9.2
torchinfo==1.8.0 